lncRNA_CIBCB2025-main/requirements.txt
alembic==1.13.2
aniso8601==9.0.1
annotated-types==0.7.0
asttokens==2.4.1
autorank==1.2.1
baycomp==1.0.3
bitarray==2.9.2
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.2.1
cycler==0.12.1
databricks-sdk==0.30.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distlib==0.3.8
docker==7.1.0
entrypoints==0.4
executing==2.0.1
filelock==3.15.4
Flask==3.0.3
fonttools==4.53.1
fpdf2==2.7.9
fsspec==2025.3.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.33.0
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
gunicorn==22.0.0
idna==3.7
imbalanced-learn==0.12.3
imblearn==0.0
importlib_metadata==7.2.1
ipython==8.26.0
iterative-stratification==0.1.9
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
Mako==1.3.5
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.1.post1
matplotlib-inline==0.1.7
mlflow==2.15.1
mlflow-skinny==2.15.1
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opentelemetry-api==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
packaging==24.1
pandas==2.2.2
parso==0.8.4
patsy==0.5.6
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==5.27.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==15.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycaleva==0.8.2
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
querystring-parser==1.2.4
requests==2.32.3
rsa==4.9
scikit-learn==1.5.1
scipy==1.14.0
setuptools==72.2.0
six==1.16.0
sklearn-compat==0.1.3
smmap==5.0.1
SQLAlchemy==2.0.32
sqlparse==0.5.1
stack-data==0.6.3
statsmodels==0.14.2
sympy==1.13.1
threadpoolctl==3.5.0
torchdata==0.9.0
tqdm==4.66.5
traitlets==5.14.3
triton==3.2.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.3
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.16.0
zipp==3.20.0
